owner check for delete

+++ questions/removeCourse.php
@@ -15,7 +15,7 @@ if (!isset($_SESSION['uid'])) {
             ->where('course',$courseName)
             ->find_one();
 
-    if ($_SESSION['type'] != 'Student'){
+    if ($_SESSION['type'] != 'Student' && $cuid->owner_uid == $uid){
 
       $cuid = $cuid->delete();
     }
